cleanup transfer function and move to separate file
